backend/requirements.txt
fastapi
uvicorn
requests
numpy
python-dotenv